backend/requirements.txt
Django==6.0.3
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
python-dotenv==1.0.1
Pillow==10.3.0
gunicorn==22.0.0
asgiref==3.8.1
sqlparse==0.5.0
typing-extensions==4.11.0
